Read skip time as integer and make sure to hande string values as well

+++ addons/binding/org.openhab.binding.allplay/src/main/java/org/openhab/binding/allplay/handler/AllPlayHandler.java
@@ -296,9 +296,9 @@ public class AllPlayHandler extends BaseThingHandler
             }
         } else if (command instanceof RewindFastforwardType) {
             if (command == RewindFastforwardType.FASTFORWARD) {
-                changeTrackPosition((long) (bindingProperties.getFastForwardSkipTimeInSec() * 1000));
+                changeTrackPosition(bindingProperties.getFastForwardSkipTimeInSec() * 1000);
             } else if (command == RewindFastforwardType.REWIND) {
-                changeTrackPosition((long) (-bindingProperties.getRewindSkipTimeInSec() * 1000));
+                changeTrackPosition(-bindingProperties.getRewindSkipTimeInSec() * 1000);
             }
         } else {
             logger.warn("Unknown control command: {}", command);
